Completed maintenance workflow

+++ src/main/java/org/twi/imageshare/web/controllers/MaintenanceController.java
@@ -11,20 +11,20 @@ import net.tanesha.recaptcha.ReCaptchaResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 import org.twi.imageshare.services.ImageService;
 import org.twi.imageshare.web.controllers.params.JsonResponse;
 
 @Controller
-@SessionAttributes("isAuthenticated")
 @RequestMapping("/maintenance")
 public class MaintenanceController {
+	
+	private static final String EMAIL = "";
+	private static final String PASSWORD = "3444c0211ae0b5275bc6a7980f9d7ac4728bb6cbc75982e0026b6fc0507a4ab8";
 	
 	@Autowired
 	private ApplicationContext context;
@@ -36,13 +36,13 @@ public class MaintenanceController {
 	private ReCaptcha recaptcha;
 
 	@RequestMapping(method = RequestMethod.GET)
-	public ModelAndView indexView() {
+	public ModelAndView indexView(HttpServletRequest paramHttpServletRequest) {
 		ModelAndView modelAndView = new ModelAndView();
-		if (isAuthenticated()) {
+		if (isAuthenticated(paramHttpServletRequest)) {
+			modelAndView.addObject("stats", imageService.getDBStats());
 			modelAndView.setViewName("maintenancePage");
 		} else {
 			modelAndView.setViewName("maintenanceAuthPage");
-			modelAndView.addObject("stats", imageService.getDBStats());
 		}
 		return modelAndView;
 	}
@@ -58,34 +58,39 @@ public class MaintenanceController {
 	    ReCaptchaResponse reCaptchaResponse = recaptcha.checkAnswer(remoteAddr, challenge, response);
 	    if (!reCaptchaResponse.isValid()) {
 	    	jsonResponse.setCaptchaError(context.getMessage("error.bad.captcha", null, Locale.getDefault()));
+	    	return jsonResponse;
 	    }
-	    
-		if (credentialsValid()) {
+		if (credentialsValid(email, password)) {
 			jsonResponse.setResponse(paramHttpServletRequest.getRequestURL());
-			
-//			modelAndView.addObject("isAuthenticated", true);  
-//			modelAndView.addObject("stats", imageService.getDBStats());
-//			modelAndView.setViewName("maintenancePage");
+			paramHttpServletRequest.getSession().setAttribute("isAuthenticated", true);
 		} else {
 	    	jsonResponse.appendError(context.getMessage("error.bad.credentials", null, Locale.getDefault()));
 		}
 		return jsonResponse;
 	}
 	
-	private boolean credentialsValid() {
-		return true;
+	private boolean credentialsValid(String email, String password) {
+		return email.toUpperCase().equals(EMAIL.toUpperCase()) && password.equals(PASSWORD);
 	}
 	
-	private boolean isAuthenticated() {
+	private boolean isAuthenticated(HttpServletRequest paramHttpServletRequest) {
+		Object result = paramHttpServletRequest.getSession().getAttribute("isAuthenticated");
+		if (result != null) {
+			return (Boolean)result;
+		}
 		return false;
 	}
 	
 	@RequestMapping(value = "/clearDB", method = RequestMethod.POST)
-	public @ResponseBody JsonResponse clearDB(ModelMap model) {
+	public @ResponseBody JsonResponse clearDB(HttpServletRequest paramHttpServletRequest) {
 		JsonResponse response = new JsonResponse();
 		try {
-			imageService.removeAll();
-			response.setSuccess(context.getMessage("app.maintenance.db.clear.success", null, Locale.getDefault()));
+			if (isAuthenticated(paramHttpServletRequest)) {
+				imageService.removeAll();
+				response.setSuccess(context.getMessage("app.maintenance.db.clear.success", null, Locale.getDefault()));
+			} else {
+				throw new Exception();
+			}			
 		} catch (Exception e) {
 			response.appendError(context.getMessage("app.maintenance.db.clear.error", null, Locale.getDefault()));
 		}
